Move azure board settings type to it's own file
