+++ src/issue-filing/services/azure-boards/azure-boards-issue-filing-service.tsx
@@ -5,21 +5,15 @@ import { IDropdownOption } from 'office-ui-fabric-react/lib/Dropdown';
 
 import { createSettingsGetter } from '../../common/create-settings-getter';
 import { IssueFilingService } from '../../types/issue-filing-service';
+import { AzureBoardsIssueDetailField, AzureBoardsIssueFilingSettings } from './azure-boards-issue-filing-settings';
 import { AzureBoardsSettingsForm } from './azure-boards-settings-form';
 import { azureBoardsIssueFilingUrlProvider } from './create-azure-boards-issue-filing-url';
 
 const AzureBoardsIssueFilingServiceKey = 'azureBoards';
 
-export type AzureBoardsIssueDetailField = 'reproSteps' | 'description';
-
 export interface AzureBoardsIssueDetailLocationDropdownOption extends IDropdownOption {
     key: AzureBoardsIssueDetailField;
 }
-
-export type AzureBoardsIssueFilingSettings = {
-    projectURL: string;
-    issueDetailsField: AzureBoardsIssueDetailField;
-};
 
 function buildStoreData(projectURL: string, issueDetailsField: AzureBoardsIssueDetailField): AzureBoardsIssueFilingSettings {
     return {
