+++ src/issue-filing/services/azure-boards/azure-boards-issue-filing-settings.tsx
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+export type AzureBoardsIssueDetailField = 'reproSteps' | 'description';
+export type AzureBoardsIssueFilingSettings = {
+    projectURL: string;
+    issueDetailsField: AzureBoardsIssueDetailField;
+};
