+++ src/tests/unit/tests/issue-filing/services/azure-boards/azure-boards-settings-form.test.tsx
@@ -8,10 +8,8 @@ import { IMock, Mock, Times } from 'typemoq';
 
 import { SettingsDeps } from '../../../../../../DetailsView/components/settings-panel/settings/settings-props';
 import { OnPropertyUpdateCallback } from '../../../../../../issue-filing/components/issue-filing-settings-container';
-import {
-    AzureBoardsIssueFilingService,
-    AzureBoardsIssueFilingSettings,
-} from '../../../../../../issue-filing/services/azure-boards/azure-boards-issue-filing-service';
+import { AzureBoardsIssueFilingService } from '../../../../../../issue-filing/services/azure-boards/azure-boards-issue-filing-service';
+import { AzureBoardsIssueFilingSettings } from '../../../../../../issue-filing/services/azure-boards/azure-boards-issue-filing-settings';
 import { AzureBoardsSettingsForm } from '../../../../../../issue-filing/services/azure-boards/azure-boards-settings-form';
 import { SettingsFormProps } from '../../../../../../issue-filing/types/settings-form-props';
 import { EventStubFactory } from '../../../../common/event-stub-factory';
